+++ lib/scroll/custom_scroll_bar.dart
@@ -16,16 +16,17 @@ class CustomScrollBar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    if (this.scrollbarStyle?.thumbColor != null) {
-      return RawScrollbar(
-        controller: this.controller,
-        isAlwaysShown: this.scrollbarStyle.isAlwaysShown,
-        thickness: this.scrollbarStyle.thickness,
-        radius: this.scrollbarStyle.radius,
-        thumbColor: this.scrollbarStyle.thumbColor,
-        child: this.child,
-      );
-    }
+    ///wait until [RawScollbar] stable or use beta version.
+    // if (this.scrollbarStyle?.thumbColor != null) {
+    //   return RawScrollbar(
+    //     controller: this.controller,
+    //     isAlwaysShown: this.scrollbarStyle.isAlwaysShown,
+    //     thickness: this.scrollbarStyle.thickness,
+    //     radius: this.scrollbarStyle.radius,
+    //     thumbColor: this.scrollbarStyle.thumbColor,
+    //     child: this.child,
+    //   );
+    // }
 
     return Scrollbar(
       controller: this.controller,